- moved the functions out in a separate baseclass that can be extended
